requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
auth0-python==4.10.0
babel==2.17.0
backoff==2.2.1
backrefs==5.8
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitarray==3.4.3
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
ckzg==2.1.1
click==8.1.7
cloudpathlib==0.21.1
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
crewai==0.130.0
crewai-tools==0.47.1
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.10
embedchain==0.1.128
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.51.0
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.3
grpcio==1.72.1
gTTS==2.5.4
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hexbytes==1.3.1
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.9.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
json5==0.12.0
json_repair==0.47.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.10.0
kiwisolver==1.4.5
kubernetes==33.1.0
lancedb==0.24.0
langchain==0.3.26
langchain-cohere==0.3.5
langchain-community==0.3.26
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
libclang==18.1.1
litellm==1.72.0
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mem0ai==0.1.110
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.5.0
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==1.27.1
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
oscrypto==1.3.0
overrides==7.7.0
packaging==24.0
paginate==0.5.7
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.16.2
pillow==10.3.0
platformdirs==4.3.8
plyer==2.1.0
portalocker==2.10.1
posthog==3.25.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pymdown-extensions==10.16
PyMuPDF==1.26.1
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==5.6.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.402
pysbd==0.3.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytube==15.0.0
pytz==2024.1
pyunormalize==16.0.0
pyvis==0.3.2
pywin32==310
PyYAML==6.0.2
pyyaml_env_tag==1.1
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.23.1
rsa==4.9.1
schema==0.7.7
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
shellingham==1.5.0.post1
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.42.2
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.6.3
typer==0.9.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
uv==0.7.13
uvicorn==0.34.3
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.1
web3==7.12.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xhtml2pdf==0.2.17
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0

